requirements.txt
# Core Dependencies
streamlit==1.32.0
ollama==0.2.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
lxml==5.1.0

# PDF and Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.11.0
pypdf==4.1.0

# Data Processing
pandas==2.2.1
python-dotenv==1.0.1

# Utilities
tqdm==4.66.2